Marconi.Core.Class

Description

Base typeclasses to index events.

See Marconi.Core for documentation.

Documentation

class Monad m => IsIndex m event indexer where

The base class of an indexer, providing its key functionalities: indexing events and handling rollbacks.

indexer the indexer implementation type

event the indexed events

m the monad in which our indexer operates

Rules:

Rollback to last indexed point do nothing: rollback p <=< index (Timed p evt) === index (Timed p evt);

Rollback eliminate intermediate indexing (precondition: none of the point in xs are equal to p): rollback p <=< indexAll events <=< index (Timed p evt) === index (Timed p evt);

Rollback is idempotent: rollback p <=< rollback p === rollback p.

Minimal complete definition

index, rollback, setLastStablePoint

Methods

index :: Eq (Point event) => Timed (Point event) (Maybe event) -> indexer event -> m (indexer event)

index an event at a given point in time

indexAll :: (Eq (Point event), Traversable f) => f (Timed (Point event) (Maybe event)) -> indexer event -> m (indexer event)

Index a bunch of event, associated to their point in time, in an indexer

The events must be sorted in ascending order (the most recent first)

indexAllDescending :: (Eq (Point event), Traversable f) => f (Timed (Point event) (Maybe event)) -> indexer event -> m (indexer event)

Index a bunch of event, associated to their point in time, in an indexer

The events must be sorted in descending order (the most recent first)

rollback :: Ord (Point event) => Point event -> indexer event -> m (indexer event)

Rollback to a previous point

setLastStablePoint :: Ord (Point event) => Point event -> indexer event -> m (indexer event)

Set the last stable point known by the indexer. Note that the last stable point should always increase, so implementation should usually filter out lower stable points.

indexIfJust

Arguments

:: IsIndex m event indexer

=> (Timed (Point event) event -> indexer event -> m (indexer event))
the function that adds the event

-> (Point event -> indexer event -> m (indexer event))
the function that update the last point, called in any case

-> Timed (Point event) (Maybe event)

-> indexer event

-> m (indexer event)

If the event is Nothing, just updat the last sync event. Otherwise, store the event and update the last sync event.

indexEither :: (IsIndex (ExceptT err m) event indexer, Eq (Point event)) => Timed (Point event) (Maybe event) -> indexer event -> m (Either err (indexer event))

Like index, but internalise its error in the result.

It's useful when you don't want to internalise the error in the monad stack to handle it explicitly, it's often used when we target IO as we don't want to mess with IOException.

indexAllEither :: (IsIndex (ExceptT err m) event indexer, Traversable f, Ord (Point event)) => f (Timed (Point event) (Maybe event)) -> indexer event -> m (Either err (indexer event))

Like indexAll, but internalise the error in the result.

indexAllDescendingEither :: (IsIndex (ExceptT err m) event indexer, Traversable f, Ord (Point event)) => f (Timed (Point event) (Maybe event)) -> indexer event -> m (Either err (indexer event))

Like indexAllDescending, but internalise the error in the result.

class HasGenesis (Point event) => Resetable m event indexer where

We can reset an indexer, clearing all its content

indexer is the indexer implementation type

event the indexer events

m the monad in which our indexer operates

Methods

reset :: indexer event -> m (indexer event)

class Queryable m event query indexer where

The indexer can answer a Query to produce the corresponding result of that query.

indexer is the indexer implementation type

event the indexer events

query the type of query we want to answer

m the monad in which our indexer operates

Methods

query :: Ord (Point event) => Point event -> query -> indexer event -> m (Result query)

Query an indexer at a given point in time It can be read as: "With the knowledge you have at that point in time, what is your answer to this query?"

queryEither :: Queryable (ExceptT (QueryError query) m) event query indexer => Ord (Point event) => Point event -> query -> indexer event -> m (Either (QueryError query) (Result query))

Like 'queryEither, but internalise QueryError in the result.

queryLatest :: Queryable m event query indexer => IsSync m event indexer => MonadError (QueryError query) m => Ord (Point event) => query -> indexer event -> m (Result query)

Like query, but use the latest point of the indexer instead of a provided one

queryLatestEither :: Queryable (ExceptT (QueryError query) m) event query indexer => IsSync m event indexer => Monad m => Ord (Point event) => query -> indexer event -> m (Either (QueryError query) (Result query))

Like queryEither, but use the latest point of the indexer instead of a provided one

class AppendResult m event query indexer where

The indexer can take a result and complete it with its events

Methods

appendResult :: Ord (Point event) => Point event -> query -> indexer event -> m (Result query) -> m (Result query)

class Closeable m indexer where

We know how to close an indexer

Methods

close :: indexer event -> m ()

class Functor m => IsSync m event indexer where

We know how far an indexer went in the indexation of events.

Methods

lastStablePoint :: indexer event -> m (Point event)

Get the latest stable point known by the indexer

lastSyncPoint :: indexer event -> m (Point event)

Last sync point of the indexer.

isAheadOfSync :: (Ord (Point event), IsSync m event indexer) => Point event -> indexer event -> m Bool

Check if the given point is ahead of the last syncPoint of an indexer

class HasGenesis t where

Any type that has a starting point

Methods

genesis :: t