Marconi.Core.Indexer.LastEventIndexer

Description

An indexer that tracks the last event received from the source.

On rollback we can either decide to remove the last event or to keep the last one we received.

Documentation

data LastEventIndexer event

An indexer that only track the latest value of an event On rollback, we either delete the value or keep the one provided before the rollback, depending on the provided configuration.

Constructors

LastEventIndexer (LastEventConfig event) (LastEventState event)

mkLastEventIndexer :: (HasGenesis (Point event), MonadIO m, MonadError IndexerError m) => LastEventConfig event -> m (LastEventIndexer event)

data LastEventConfig event

Constructors

LastEventConfig FilePath Bool (Point event -> ByteString) (ByteString -> Either Text (Point event)) (event -> ByteString) (ByteString -> Either Text event) Word

data GetLastQuery a

Datatype used to query the stored value of a LastEventIndexer

Constructors

GetLastQuery