Marconi.Core.Indexer.LastPointIndexer

Description

On-disk indexer backed by a sqlite database.

See Marconi.Core for documentation.

Documentation

data LastPointIndexer event

LastPointIndexer. An indexer that does nothing except keeping track of the last point. While it may sound useless, it can be usefull when you want to benefit of the capabilities of a transformer.

lastPointIndexer :: HasGenesis (Point event) => LastPointIndexer event

A smart constructor for LastPointIndexer