Marconi.Core.Preprocessor.Resume

Description

Resume indexing after a previous run: this preprocessor drain events up to the last stable event indexed by an indexer

Documentation

type Resume m event = Preprocessor m (Point event) event event

A type alias for a preprocessor that allows resuming. We don't use a newtype to ease preprocessors composition.

withResume

Arguments

:: (MonadError IndexerError m, Ord (Point event))

=> Point event
The last point of an indexer.

-> Resume m event