Marconi.Core.Query

Description

A set of queries that can be implemented by any indexer

See Marconi.Core for documentation.

Documentation

data EventAtQuery event

Get the event stored by the indexer at a given point in time

Constructors

EventAtQuery

newtype EventsMatchingQuery event

Query an indexer to find all events that match a given predicate

The result should return the most recent first

Constructors

EventsMatchingQuery (event -> Maybe event)

allEvents :: EventsMatchingQuery event

Get all the events that are stored in the indexer

newtype LatestEventsQuery event

Get the nbOfEvents last non empty events from the indexer before the point given in the query

Constructors

LatestEventsQuery Word

latestEvent :: LatestEventsQuery event

Get the latest non empty event before the point given in the query

newtype EventsFromQuery event

Get the non empty events from the given point (excluded) to the one of the query (included)

Constructors

EventsFromQuery (Point event)

data Stability a

Represents whether an event is considered to stable or not.

Constructors

Stable a

Volatile a

isStable :: Stability a -> Bool

Tests for stability

calcStability :: Ord point => (event -> point) -> point -> event -> Stability event

Helper function to wrap an event in Stability based on the last stable point and a given way to get a point.

Asks: Is the provided point less than (from a time before) or equal to (from the time of) the last stable point?

newtype WithStability query

A wrapper representing the demand that the result of a query should come with Stability data

Constructors

WithStability

Fields

unWithStability :: query

withStabilityM

Arguments

:: (Result query ~ f event, Result (WithStability query) ~ f (Stability event), MonadError (QueryError (WithStability query)) m, Traversable f)

=> (event -> Stability event)
A function to calculate Stability of an event

-> ExceptT (QueryError query) m (f event)
In ExceptT action whose result is an event

-> m (f (Stability event))

Convert a non-Stability supporting instance to a Stability supporting instance.

withStability

Arguments

:: forall event m indexer f. (Monad m, Ord (Point event), IsSync m event indexer, Traversable f)

=> indexer event
An indexer

-> f (Timed (Point event) event)
A traversable of query results

-> m (f (Stability (Timed (Point event) event)))

Given an indexer and some traversable of timed query results, calculate the stability of all the query results.

withStabilityAt

Arguments

:: forall result m event indexer f. (Monad m, Ord (Point event), IsSync m event indexer, Applicative f)

=> indexer event
An indexer

-> Point event
A specific point to compare against the last stable point

-> f result
A traversable of query results

-> m (f (Stability result))

Given an indexer, a Point and some traversable of query results, calculate the stability of all the query results.

queryErrorWithStability :: (Result (WithStability query) ~ f (Stability event), Result query ~ f event) => (f event -> f (Stability event)) -> QueryError query -> QueryError (WithStability query)