Marconi.Core.TracedStorable

State

data Config

memoryBufferSize :: Iso' Config Int

data State h

handle :: forall h. Lens' (State h) h

config :: forall h. Lens' (State h) Config

emptyState :: PrimMonad (StorableMonad h) => Int -> h -> StorableMonad h (State h)

data Storage h

storage :: forall h. Lens' (State h) (Storage h)

events :: forall h h. Lens (Storage h) (Storage h) (MVector (PrimState (StorableMonad h)) (SyntheticEvent (StorableEvent h) (StorablePoint h))) (MVector (PrimState (StorableMonad h)) (SyntheticEvent (StorableEvent h) (StorablePoint h)))

cursor :: forall h. Lens' (Storage h) Int

getMemoryEvents :: Storage h -> MVector (PrimState (StorableMonad h)) (SyntheticEvent (StorableEvent h) (StorablePoint h))

getEvents :: Buffered h => PrimMonad (StorableMonad h) => State h -> StorableMonad h [StorableEvent h]

data family StorableEvent h

type family StorablePoint h

The resume and query functionality requires a way to specify points on the chain from which we want to resume, or points up to which we want to query.

data family StorableQuery h

data family StorableResult h

type family StorableMonad h :: * -> *

data family StorableNotifications h

API

data SyntheticEvent e p

Constructors

Event !e

Synthetic !p

data ControlNotification pt n

Constructors

CNRollForward !pt

CNRollBack !pt

CNApplication !n

class Buffered h where

Methods

persistToStorage :: Foldable f => AppTracer h -> f (StorableEvent h) -> h -> StorableMonad h h

This function persists the memory/buffer events to disk when the memory buffer is filled.

getStoredEvents :: h -> StorableMonad h [StorableEvent h]

class Queryable h where

Methods

queryStorage :: Foldable f => AppTracer h -> StorablePoint h -> f (StorableEvent h) -> h -> StorableQuery h -> StorableMonad h (StorableResult h)

class Resumable h where

Methods

resumeFromStorage :: AppTracer h -> h -> StorableMonad h [StorablePoint h]

class Rewindable h where

Methods

rewindStorage :: AppTracer h -> StorablePoint h -> h -> StorableMonad h (Maybe h)

class HasPoint e p where

Methods

getPoint :: e -> p

syntheticPoint :: HasPoint e p => SyntheticEvent e p -> p

foldEvents :: forall f h. Foldable f => f (SyntheticEvent (StorableEvent h) (StorablePoint h)) -> [StorableEvent h]

insert :: Buffered h => PrimMonad (StorableMonad h) => HasPoint (StorableEvent h) (StorablePoint h) => ControlTracer h -> StorableEvent h -> State h -> StorableMonad h (State h)

checkpoint :: Buffered h => PrimMonad (StorableMonad h) => AppTracer h -> StorablePoint h -> State h -> StorableMonad h (State h)

insertMany :: Foldable f => Buffered h => PrimMonad (StorableMonad h) => HasPoint (StorableEvent h) (StorablePoint h) => ControlTracer h -> f (StorableEvent h) -> State h -> StorableMonad h (State h)

rewind :: forall h. Rewindable h => PrimMonad (StorableMonad h) => Ord (StorablePoint h) => HasPoint (StorableEvent h) (StorablePoint h) => ControlTracer h -> StorablePoint h -> State h -> StorableMonad h (Maybe (State h))

resume :: Resumable h => AppTracer h -> State h -> StorableMonad h [StorablePoint h]

query :: Queryable h => PrimMonad (StorableMonad h) => AppTracer h -> StorablePoint h -> State h -> StorableQuery h -> StorableMonad h (StorableResult h)