Marconi.Core.Transformer.Class

Description

Typeclass for indexer transformers,

It allows the generalisation of properties amongst all indexer transformers.

Documentation

class IndexerTrans t where

An indexer transformer: it adds a configurable capability to a tranformer

Methods

unwrap :: Lens' (t indexer event) (indexer event)

Unwrap the underlying indexer

class IndexerMapTrans t where

An indexer transformer: it adds a configurable capability to a tranformer

This one allow also the transformation of the event, contrary to IndexerTrans.

Methods

unwrapMap :: Lens' (t indexer output event) (indexer output)

Unwrap the underlying indexer