Marconi.Core.Transformer.WithAction

Description

A transformer that fans events into an arbitrary action.

The configuration is a function Timed (Point event) event -> IO ()

See Marconi.Core.Transformer.WithStream.Socket and Marconi.Core.Transformer.WithStream.TBQueue for example usages.

Documentation

newtype WithAction indexer event

An indexer which runs an action on events during indexing

Constructors

WithAction

Fields

_withActionWrapper :: IndexTransformer WithActionConfig indexer event

withActionWrapper :: forall indexer event indexer event. Iso (WithAction indexer event) (WithAction indexer event) (IndexTransformer WithActionConfig indexer event) (IndexTransformer WithActionConfig indexer event)

newtype WithActionConfig event

Configuration is a function Timed (Point event) event -> IO ()

Constructors

WithActionConfig

Fields

_withActionConfigAction :: Timed (Point event) event -> IO ()

withActionConfigAction :: forall event event. Iso (WithActionConfig event) (WithActionConfig event) (Timed (Point event) event -> IO ()) (Timed (Point event) event -> IO ())