Marconi.Core.Transformer.WithCache

Description

A transformer that cache result of some queries

See Marconi.Core for documentation.

Documentation

data WithCache query indexer event

Setup a cache for some requests.

The cache is active only for the latest Point. As a consequence, using WithCache is more effective on top of the on-disk part of a MixedIndexer, or any other part of an indexer that has a relatively stable sync point.

withCache

Arguments

:: Ord query

=> (Timed (Point event) (Maybe event) -> Result query -> Result query)
The cache update function

-> indexer event

-> WithCache query indexer event

A smart constructor for WithCache. The cache starts empty, you can populate it with addCacheFor

addCacheFor :: Queryable (ExceptT (QueryError query) m) event query indexer => IsSync (ExceptT (QueryError query) m) event indexer => HasCacheConfig query indexer => Monad m => MonadError IndexerError m => Ord query => Ord (Point event) => query -> indexer event -> m (indexer event)

Add a cache for a specific query.

When added, the WithCache queries the underlying indexer to populate the cache for this query.

If you want to add several indexers at the same time, use traverse.

class HasCacheConfig query indexer where

Methods

cache :: Lens' (indexer event) (Map query (Result query))

Access to the cache