Marconi.Core.Transformer.WithCatchup

Description

The catch-up mechanism buffered events to pass them in batch to an indexer, to reduce the time needed for an indexer to catch-up with the tip.

Once we are close enough to the tip of the chain, WithCatchup deactivate itself and pass directly the blocks to the indexer.

When you use several workers and coordinators, you may want to put the catchup on the final indexers and not on the coordinators to improve performances.

Documentation

data WithCatchup indexer event

withCatchup

Arguments

:: (Point event -> event -> Word64)
The distance function

-> CatchupConfig
Configure how many element we put in a batch and until when we use it

-> indexer event
the underlying indexer

-> WithCatchup indexer event

A smart constructor for WithCatchup

data CatchupConfig

The visible part of the catchup configuration, it allows you to configure the size of the batch and to control when the batch mecanism stops

Constructors

CatchupConfig Word64 Word64 (Maybe (CatchupEvent -> IO ()))

mkCatchupConfig :: Word64 -> Word64 -> CatchupConfig

configCatchupEventHook :: Lens' CatchupConfig (Maybe (CatchupEvent -> IO ()))

class HasCatchupConfig indexer where

A typeclass that allows an indexer with a WitchCatchup transformer to configure the behaviour of this transformer

Methods

catchupBypassDistance :: Lens' (indexer event) Word64

catchupBatchSize :: Lens' (indexer event) Word64

catchupEventHook :: Lens' (indexer event) (Maybe (CatchupEvent -> IO ()))

data CatchupEvent

Constructors

Synced