Marconi.Core.Transformer.WithDelay

Description

A transformer that delay insertion of events into an indexer.

See Marconi.Core for documentation.

Documentation

data WithDelay indexer event

When indexing computation is expensive, you may want to delay it to avoid expensive rollback WithDelay buffers events before sending them to the underlying indexer. Buffered events are sent when the buffers overflows.

An indexer wrapped in WithDelay won't interact nicely with a coordinator at the moment, as WithDelay acts as it's processing an event while it only postpones the processing.

As a consequence, WithDelay is preferably used at the top of the hierarchy.

withDelay

Arguments

:: Word
capacity

-> indexer event

-> WithDelay indexer event

A smart constructor for WithDelay

class HasDelayConfig indexer where

Provide accces to the delay size of a WithDelay transformer. The provided instance ensure that access is granted even if other indexer transformers are used on top of this one

Methods

delayCapacity :: Lens' (indexer event) Word