Marconi.Core.Transformer.WithFold

Description

A transformer that allows an indexer to fold incoming events: each new event is the result of a computation between the last result in the indexer and the incoming event.

See Marconi.Core for documentation.

Documentation

data WithFold m indexer output input

WithFold fold incoming event to produce an output.

withFold :: output -> (indexer output -> m (Maybe output)) -> (output -> input -> m output) -> indexer output -> WithFold m indexer output input

A smart constructor for WithFold

withFoldPure :: Applicative m => output -> (indexer output -> m (Maybe output)) -> (output -> input -> output) -> indexer output -> WithFold m indexer output input

A smart constructor for WithFold

withFoldMap :: (Monoid output, Applicative m) => (indexer output -> m (Maybe output)) -> (input -> output) -> indexer output -> WithFold m indexer output input

A smart constructor for WithFold

getLastEventAtQueryValue :: (IsSync m event indexer, HasGenesis (Point event), Queryable (ExceptT (QueryError (EventAtQuery a)) m) event (EventAtQuery a) indexer, Ord (Point event), MonadError IndexerError m) => indexer event -> m (Maybe a)

class HasFold m input output indexer where

There are few scenarios where you want to modify the fold function but it may happen

Methods

fold :: Lens' (indexer input) (output -> input -> m output)