Marconi.Core.Transformer.WithPruning

Description

A transformer that prune the content of an indexer

See Marconi.Core for documentation.

Documentation

class Prunable m event indexer where

The indexer can prune old data. The main purpose is to speed up query processing. If the indexer is Rollbackable and Prunable, it can't rollback behind the pruningPoint, the idea is to call prune on points that can't be rollbacked anymore.

indexer is the indexer implementation type

desc the descriptor of the indexer, fixing the Point types

m the monad in which our indexer operates

Methods

prune :: Ord (Point event) => Point event -> indexer event -> m (indexer event)

pruningPoint :: indexer event -> m (Maybe (Point event))

pruneVia :: (Functor m, Prunable m event indexer, Ord (Point event)) => Lens' s (indexer event) -> Point event -> s -> m s

Helper to implement the prune functon of Prunable when we use a wrapper. Unfortunately, as m must have a functor instance, we can't use deriving via directly.

pruningPointVia :: Prunable m event indexer => Getter s (indexer event) -> s -> m (Maybe (Point event))

Helper to implement the pruningPoint functon of Prunable when we use a wrapper. Unfortunately, as m must have a functor instance, we can't use deriving via directly.

data WithPruning indexer event

WithPruning control when we should prune an indexer

The main purpose is to optimize storage for events that can't be rollbacked anymore.

In some contexts, once you know that an event can't be rollback anymore, your indexer may not need it, or may process it's information to make it irrelevant. In this case, you may want to prune the stored events.

withPruning

Arguments

:: Word
how far can a rollback go

-> Word
once we have enough events, how often do we prune

-> indexer event

-> WithPruning indexer event

nextPruning :: Lens' (WithPruning indexer event) (Seq (Point event))

stepsBeforeNext :: Lens' (WithPruning indexer event) Word

currentDepth :: Lens' (WithPruning indexer event) Word

class HasPruningConfig indexer where

Methods

securityParam :: Lens' (indexer event) Word

pruneEvery :: Lens' (indexer event) Word