Marconi.Core.Transformer.WithStream

Documentation

withStream :: Streamable a r => (Timed (Point event) event -> r) -> a -> indexer event -> WithAction indexer event

A smart constructor for WithAction, using any streamable.

Be aware: depending on the implementation you choose, this can lead to the indexer blocking, given a slow (or non-existent) consumer!