Marconi.Core.Transformer.WithTracer

Description

A transformer that add tracing capability to an indexer.

See Marconi.Core for documentation.

Event

data IndexerEvent point

Event available for the tracer

Constructors

IndexerIsStarting

IndexerFailed IndexerError

IndexerStarted

IndexerIndexes point

IndexerHasIndexed

IndexerRollbackTo point

IndexerHasRollbackedTo

IndexerIsClosing

IndexerClosed

Trace

data WithTrace m indexer event

A tracer modifier that adds tracing to an existing indexer

withTrace :: Applicative m => Trace m (IndexerEvent (Point event)) -> indexer event -> WithTrace m indexer event

A smart constructor for WithTrace

withTraceM :: MonadIO m => Trace m (IndexerEvent (Point event)) -> m (indexer event) -> m (WithTrace m indexer event)

A monadic smart constructor for WithTrace

class HasTraceConfig m event indexer where

It gives access to the tracer. The provided instances allows access to the tracer event below an indexer transformer.

Methods

trace :: Lens' (indexer event) (Trace m (IndexerEvent (Point event)))

Tracer

data WithTracer m indexer event

A tracer modifier that adds tracing to an existing indexer

withTracer :: Applicative m => Tracer m (IndexerEvent (Point event)) -> indexer event -> WithTracer m indexer event

A smart constructor for WithTracer

withTracerM :: Applicative m => Tracer m (IndexerEvent (Point event)) -> m (indexer event) -> m (WithTracer m indexer event)

A monadic smart constructor for WithTracer

class HasTracerConfig m event indexer where

It gives access to the tracer. The provided instances allows access to the tracer event below an indexer transformer.

Methods

tracer :: Lens' (indexer event) (Tracer m (IndexerEvent (Point event)))