Marconi.Core.Transformer.WithTransform

Description

A transformer that map the events in an indexer.

See Marconi.Core for documentation.

Documentation

data WithTransform indexer output input

WithTransform is an indexer transformer that map the event type of an indexer

config :: forall indexer output input input. Lens (WithTransform indexer output input) (WithTransform indexer output input) (TransformConfig output input) (TransformConfig output input)

withTransform :: (Point input -> Point output) -> (input -> Maybe output) -> indexer output -> WithTransform indexer output input

A smart constructor for WithTransform

class HasTransformConfig input output indexer where

Provide accces to the tranformation function of a WithTransform transformer. The provided instance ensure that access is granted even if other indexer transformers are used on top of this one

Methods

transformEvent :: Lens' (indexer input) (input -> Maybe output)