Marconi.Core.Type

Description

Base types to index events.

See Marconi.Core for documentation.

Types and type families

type family Point event

A point in time, the concrete type of a point is now derived from an indexer event, instead of an event. The reason is that you may not want to always carry around a point when you manipulate an event.

data ProcessedInput point event

The different types of input event that should be handled by an indexer used to map the chain incoming events to something that an indexer should be able to digest.

Constructors

Rollback point
A rollback happen and indexers need to go back to the given point in time

Index (Timed point (Maybe event))
A new event has to be indexed

IndexAllDescending (NonEmpty (Timed point (Maybe event)))
A new event has to be indexed

StableAt point
Inform the indexer of the latest stable point reached

Stop
Processing stops

type family Result query

A Result is a data family for query descriptor.

A query is tied to an indexer by a typeclass, this design choice has two main reasons: * we want to be able to define different query for the same indexer (eg. we may want to define two distinct query types for an utxo indexer: one to ge all the utxo for a given address, another one for to get all the utxos emitted at a given slot). * we want to assign a query type to different indexers.

data Timed point event

Attach an event to a point in time

Constructors

Timed point event

point :: Lens' (Timed point event) point

When was this event created

event :: Lens (Timed point a) (Timed point b) a b

A lens to get the event without its time information

Error types

data IndexerError

Error that can occur when you index events

Constructors

RollbackBehindHistory
An indexer don't have access to the history at the point that is asked

IndexerInternalError Text
The indexer did not respond

InvalidIndexer Text
The indexer is in an invalid state and can't recover

StopIndexer (Maybe Text)
Indexer has to stop as requested by the given worker

ResumingFailed Text
The indexer failed at resuming (likely due to a bug)

IndexerCloseTimeoutError
The indexer timed out while attempting to write a file before closing

OtherIndexError Text
Any other cause of failure

_RollbackBehindHistory :: Prism' IndexerError ()

_IndexerInternalError :: Prism' IndexerError Text

_InvalidIndexer :: Prism' IndexerError Text

_StopIndexer :: Prism' IndexerError (Maybe Text)

_ResumingFailed :: Prism' IndexerError Text

_IndexerCloseTimeoutError :: Prism' IndexerError ()

_OtherIndexError :: Prism' IndexerError Text

data QueryError query

Error that can occurs when you query an indexer

Constructors

AheadOfLastSync (Maybe (Result query))
The required point is ahead of the current index. The error may still provide its latest result if it make sense for the given query.

It can be useful for indexer that contains a partial knowledge and that want to pass this knowledge to another indexer to complete the query.

NotStoredAnymore
The requested point is too far in the past and has been pruned

IndexerQueryError Text
The indexer query failed

SlotNoBoundsInvalid Text
Upper or lower SlotNo bounds provided in the query are not consistent. For example, the requested point is too early to answer the query completely.

_AheadOfLastSync :: forall query query. Prism (QueryError query) (QueryError query) (Maybe (Result query)) (Maybe (Result query))

_NotStoredAnymore :: forall query. Prism' (QueryError query) ()

_IndexerQueryError :: forall query. Prism' (QueryError query) Text

_SlotNoBoundsInvalid :: forall query. Prism' (QueryError query) Text